+++ src/networkmessage.cpp
@@ -100,7 +100,7 @@ void NetworkMessage::addItem(const Item* item)
 	add<uint16_t>(it.clientId);
 
 	if (it.stackable) {
-		addByte(std::min<uint16_t>(0xFF, item->getItemCount()));
+		addByte(static_cast<uint8_t>(std::min<uint16_t>(0xFF, item->getItemCount())));
 	} else if (it.isSplash() || it.isFluidContainer()) {
 		addByte(fluidMap[item->getFluidType() & 7]);
 	}
